OpenLanding: factor-out ChevronLink and Screenshots
