+++ catalog/app/website/components/Screenshots/index.js
@@ -1,0 +1,1 @@
+export default from './Screenshots'
